Added/modified some files related to the client and server sides:
Modified the .env file for the client and server.
Modified the logo and yariga.svg files in the client's assets folder.
Modified the TotalRevenue.tsx and chart.config.ts files in the client's charts folder.
Modified the Form.tsx and Profile.tsx files in the client's common folder.
Modified the agent.tsx, all-properties.tsx, home.tsx, and property-details.tsx files in the client's pages folder.
Modified the .env file for the server.